+++ packages/core/src/components/panel-stack2/panelStack.tsx
